Корректировки по TaskController и TaskService

+++ app/Services/TaskService.php
@@ -33,18 +33,21 @@ readonly class TaskService
         $this->taskRepository->createTask($data);
     }
 
-    public function updateTask(Task $task, array $data): void
+    public function updateTask(int $userId, Task $task, array $data): void
     {
+        abort_if(!$this->taskRepository->hasTaskAccess($userId, $task), 403);
         $task->update($data);
     }
 
-    public function toggleTask(Task $task): void
+    public function toggleTask(int $userId, Task $task): void
     {
+        abort_if(!$this->taskRepository->hasTaskAccess($userId, $task), 403);
         $task->update(['is_done' => !$task->is_done]);
     }
 
-    public function deleteTask(Task $task): void
+    public function deleteTask(int $userId, Task $task): void
     {
+        abort_if(!$this->taskRepository->hasTaskAccess($userId, $task), 403);
         $task->delete();
     }
 }
